--- tests/units/v2/test_location_name_repository.py
@@ -1,7 +1,0 @@
-import unittest
-
-
-class TestLocationNameRepository(unittest.TestCase):
-
-    def test_get_location_name(self):
-        pass
